refactor register pages
