+++ src/pages/LoginRegister.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import StatusBar from "../components/loginRegisterPage/StatusBar";
-import SubmitForm from "../components/loginRegisterPage/SubmitForm";
+import StatusBar from "../components/specSection/StatusBar";
+import SubmitForm from "../components/specSection/SubmitForm";
 import { useParams, useNavigate } from "react-router-dom";
 
 function LoginRegister() {
